Updated gopkg.lock; Removed error code;
Added db.close to some testcases;
A detailed explanation on the 2 lines of code that produced error here:
blockchain_test.go uses GenerateUtxoMockBlockWithoutInputs which returns
a block reference. By default another testcase (TestCopyAndRevertUtxos)
was affected when block value was set in blockchain_test.go as it used
the same block reference. Thus the 2 lines of code that changed the block value was removed.