add addAccount command to cli, fix vinRules
To allow regular transactions between miners and normal accounts, we
need a way to add accounts to a wallet via private key inputs. This
commit adds the addAccount command to cli which takes a private key as
an argument and adds a matching account to the wallet. The operation is
prevented if the account already exists in the wallet.

In order to allow transactions to be sent from the miner, a change had
to be made to the vinRules function. Currently if the 'from' account
in a transaction has 100+ utxos in their pool, vinRules will force the
transaction to include 51 utxos as inputs even if the transaction could
be done with fewer. This ensures the transaction will eventually fail.
Since miner accounts tend to have large utxo pools, this is problematic.
My interpretation of the intended behaviour was that as many
utxos as possible (50) should be included in the transaction in order to
shrink the utxo pool. I therefore changed vinRules to only include
up to 50 utxos on the assumption that including 51 utxos was an error.
This allows transactions with a large utxo pool to succeed.
I also removed ErrTooManyUtxoFund as it appears unreachable.