create DIDManager contract, add new DIDs
This commit adds the ./cli createDID command. It first sends a DIDManager
contract to the blockchain if it doesn't already exist
(address stored locally). It will then generate a DID and corresponding
DID document and send them to the DIDManager contract. A contractQuery
can be sent to the DIDManager using the manager's address and the DID
as the key (in quotation marks) to retrive the document.
Created DIDs are also stored in a local wallet along with an associated
private key. This will be used for future verification.