full implementation, add/update/delete values
Using ./cli updateDID with either a did or the name of a basic DID doc
as a parameter, you can view all the values in the corresponding full
DID doc. Verification is performed with a timestamp as a signature.
All modifiable values (those not in the basic doc) are stored in a map
to allow arbitrary keys to be easily created.